drop support for go versions <= 1.7
